should be close to last of clue scroll stuff
the only thing left should be if a npc is inaccessible, which I believe
none are left.

I checked every single clue scroll and commented out ones that do not
work.

people that say "b-but muh clue scroll doesn't work" are idiots and just
aren't looking in the right spot.