ODEL PLS LOOK
okay, I added all the dialogues and such. But there are a few problems.

I can't figure out how to make it so clicking on "The Knight's Sword" in
the quest tab works
and
I can't figure out how to make it save to characters.txt
and
I can't find the ID of the object for the blurite rocks to make it so
you can mine them.

beyond that everything should be good, feel free to look through it and
check.

I'm leaving for work, so fix these if you can before I get back if you
get on.