Made changes to region music (read dec odel)
Before everytime  playmusic was called it was just creating a new class
instance everytime in the walkhandler it has an effect on how things
work i would say So i have made it create a class instance for each
player so no new instance needs to be made everytime. As well as that i
added the playmusic call to reload region which will call before the
walk handler this is specifically usefull when a region obviously
reloads and teleporting.