Highscores first pass with sqlite
Set Constants.SQL_ENABLED to true, a database file will be created when
the server starts (this can take some time if you have a lot of save
files).  It will proceed to go into every save file and save the total
level and xp to the database.  You can use ::highscores to see a list of
the top ~ 45 entries.  If you want some character files to play with
this, just msg me on steam.