Creature of Fenkenstrain Progress 2
Quest should still be disabled but added it to the QuestHandler
commented out
Added Conductor crafting. I decided to give it a separate interface from
the SilverCrafting one since the conductor is useless outside of the
quest. If the players have the conductor mould in their inventory when
using a silver bar on a furnace, it'll bring up the conductor mould
interface, otherwise it'll just default to the regular interface. Craft
1,5,X and All work.

Added some more object interactions/dialogues and cleaned up a few
things.

Leme know if there is something I need to do differently for some
reason.